python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "colorlog>=6.9.0",
    "coremltools>=8.3.0",
    "grpcio>=1.73.1",
    "grpcio-health-checking>=1.73.1",
    "numpy>=2.3.0",
    "onnx==1.16.0",
    "onnxmltools==1.13.0",
    "onnxruntime>=1.22.0",
    "pandas>=2.3.1",
    "pyarrow>=20.0.0",
    "pytest>=8.4.1",
    "tensorboard>=2.20.0",
    "torch==2.7.0",
    "grpcio-tools>=1.73.1",
    "matplotlib>=3.10.7",
    "scikit-learn>=1.7.2",
]


#[tool.uv.sources]
#torch = [{ index = "pytorch-cu128", marker = "platform_system != 'Darwin'" }]
#
#[[tool.uv.index]]
#name = "pytorch-cu128"
#url = "https://download.pytorch.org/whl/cu124"
#explicit = true

# To use group 'gh' (grace hopper) run with 'uv sync --group gh'
#[project.optional-dependencies]
#gh = [
#    "torch==2.7.0+cu128",
#    "markupsafe==2.1.3",
#]
#
#[dependency-groups]
#dev = [
#    "grpcio-tools>=1.73.1",
#]
#
#gh = [
#    "torch==2.7.0+cu128",
#    "markupsafe==2.1.3",
#]
#
#[tool.uv.sources]
#torch = [
#{ index = "torch-gh", extra = "gh" },
#]
#
#[[tool.uv.index]]
#name = "torch-gh"
#url = "https://download.pytorch.org/whl/cu128"
#


#[dependency-groups]
#dev = [
#    "grpcio-tools>=1.73.1",
#]
#normal = [
#    "torch==2.7.0",
#]
#cuda = [
#    "torch==2.7.0+cu128",
#]
#
#[tool.uv.sources]
#torch = [
##    { index = "torch-normal", extra = "normal" },
#    { index = "torch-cuda", extra = "cuda" },
#]
#
#[[tool.uv.index]]
#name = "torch-normal"
#url = "https://download.pytorch.org/whl/cpu"
#
#[[tool.uv.index]]
#name = "torch-cuda"
#url = "https://download.pytorch.org/whl/cu128"

#[project.optional-dependencies]
#dev = [
#    "grpcio-tools>=1.73.1",
#]
#gh = [
#    # Install CUDA-enabled torch ONLY on Linux
#    "torch==2.7.0+cu128; sys_platform == 'linux'",
#    "markupsafe==2.1.3", # This is likely needed for the Jinja version used by torch
#]
#

# This section is the key for uv on Grace Hopper.
# It tells uv: "When looking for the package 'torch' AND the 'gh' group is active,
# check this special index first."
#[tool.uv.sources]
#torch = [
#    { index = "torch-gh", extra = "gh" },
#]
#
## Definition for the special index.
#[[tool.uv.index]]
#name = "torch-gh"
#url = "https://download.pytorch.org/whl/cu128"